П. Р. 11

+++ src/utils/api.js
@@ -28,11 +28,11 @@ updateUserProfile({ name, about }) {
     }).then(this._checkStatus);
   }
 
-  updateUserAvatar(avatar) {
+  updateUserAvatar(data) {
     return fetch(`${this._baseUrl}/users/me/avatar`, {
       method: "PATCH",
       body: JSON.stringify({
-        avatar,
+        avatar: data.avatar,
       }),
       headers: this._headers,
     }).then(this._checkStatus);
@@ -45,13 +45,13 @@ updateUserProfile({ name, about }) {
       }).then(this._checkStatus);
     }
 
-    postCard({ title, link }) {
+    postCard(data) {
       return fetch(`${this._baseUrl}/cards`, {
         method: "POST",
         headers: this._headers,
         body: JSON.stringify({
-          name: title,
-          link: link,
+          name: data.name,
+          link: data.link,
         }),
       }).then(this._checkStatus);
     }
@@ -63,22 +63,20 @@ updateUserProfile({ name, about }) {
       }).then(this._checkStatus);
     }
 
-    _plusLike(id) {
-      return fetch(`${this._baseUrl}/cards/${id}/likes`, {
-        method: "PUT",
-        headers: this._headers,
-      }).then(this._checkStatus);
-    }
-
-    _minusLike(id) {
-      return fetch(`${this._baseUrl}/cards/${id}/likes`, {
-        method: "DELETE",
-        headers: this._headers,
-      }).then(this._checkStatus);
-    }
-
-    likeCard({ cardId, isLiked }) {
-      return isLiked ? this._minusLike(cardId) : this._plusLike(cardId);
+    likeCard(cardId, isLiked) {
+      if (isLiked) {
+        return fetch(`${this._baseUrl}/cards/${cardId}/likes`, {
+          method: 'DELETE',
+          headers: this._headers,
+        })
+          .then(this._checkStatus);
+      } else {
+        return fetch(`${this._baseUrl}/cards/${cardId}/likes`, {
+          method: 'PUT',
+          headers: this._headers,
+        })
+          .then(this._checkStatus);
+      }
     }
 }
